Create BlueStorageUtility and move related functions
In the past additional features such as copy to Azure were implemented on the settings page. These features are going to be moved out of the settings page and onto a dedicated page on the admin menu. This provides clean separation between functions allowing custom roles to better manage access to BlueStorage. This commit does not fully implement the new page but is just a first step to having the separation.