Remove const class as I don't think it is needed
I don't think there is much need for these settings. I don't think it was a good implementation strategy either. It's possible I may reimplement some of these but it would be in a whole new way and maybe use custom defined capabilities that can be used in custom roles.